Updated to make script re-entrant
If you ran the script twice on the same playlist then it would screw up the length of the play list in plex (it was resported as double the length of all the tracks).

This fix allows the script to be re-run. It will also allow you to import a playlist add a couple more songs and re-import it without having to delete from plex first (however that will mean ordering won't be the same in Plex).